+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+5.7.1
+=====
+
+*   (improvement) Return the `item` + `before` in the `Sortable` result (to integrate with the new RAD bundle).
+*   (improvement) Added `Sortable::destroy()`.
+
+
 5.7.0
 =====